Danish/Aarhus University fellowships
Not sure if that is the place where would put that fellowship, but Aarhus University (DK) is providing 12 fellowships. 
The fellowships are to support researchers from **any** scientific discipline, who are currently affected by the war in Ukraine and unable to continue their research in the light of the current situation. The fellowships will allow researchers to continue their research and collaborate with other researchers in a secure, international and interdisciplinary research environment.